QuadTree v1
Working QuadTrees!!
It's finally done, supports infinite amounts of enemies - but the AI
calculations take FOREVER... 3 enemies at a time will slow things down
considerably and (on my pc) 6 will cause visual studio to fully run out
of allocated memory :package: slow things down considerably and (on my
pc) 6 will cause visual studio to fully run out of allocated memory :P

-MUST RUN IN RELEASE MODE!!!!

+++ Game Start/QuadTree.cpp
@@ -22,10 +22,10 @@ void QuadTree::clear(){
 }
 
 void QuadTree::split(){
-	int subWidth = bounds.getSize().x / 2;
-	int subHeight = bounds.getSize().y / 2;
-	int x = bounds.getPosition().x;
-	int y = bounds.getPosition().y;
+	float subWidth = bounds.getSize().x / 2;
+	float subHeight = bounds.getSize().y / 2;
+	float x = bounds.getPosition().x;
+	float y = bounds.getPosition().y;
 
 	sf::RectangleShape rectangle(sf::Vector2f(subWidth,subHeight));
